requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-ollama
lxml
chromadb
faiss-cpu
langchain_chroma
langchain_groq
langchain_core
fastapi
uvicorn
langserve
transformers
sentence-transformers
langchain_huggingface
streamlit
Ollama
duckduckgo-search
pymupdf
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
requests